requirements.txt
selenium
webdriver_manager
requests
flask
flask-socketio
eventlet
Pillow
